+++ myems-admin/app/controllers/settings/combinedequipment/combinedequipmentmeter.controller.js
@@ -184,13 +184,34 @@ app.controller('CombinedEquipmentMeterController', function (
         });
     };
 
-    $scope.getAllCombinedEquipments();
-    $scope.getAllMeters();
-    $scope.getAllVirtualMeters();
-    $scope.getAllOfflineMeters();
+    $scope.tabInitialized = false;
+
+    $scope.initTab = function() {
+        if (!$scope.tabInitialized) {
+            $scope.tabInitialized = true;
+            $scope.getAllCombinedEquipments();
+            $scope.getAllMeters();
+            $scope.getAllVirtualMeters();
+            $scope.getAllOfflineMeters();
+        }
+    };
+
+    $scope.$on('combinedequipment.tabSelected', function(event, tabIndex) {
+        if ($scope.$parent && $scope.$parent.TAB_INDEXES && tabIndex === $scope.$parent.TAB_INDEXES.BIND_METER && !$scope.tabInitialized) {
+            $scope.initTab();
+        }
+    });
+
+    $timeout(function() {
+        if ($scope.$parent && $scope.$parent.TAB_INDEXES && $scope.$parent.activeTabIndex === $scope.$parent.TAB_INDEXES.BIND_METER && !$scope.tabInitialized) {
+            $scope.initTab();
+        }
+    }, 0);
 
     $scope.$on('handleBroadcastCombinedEquipmentChanged', function (event) {
-        $scope.getAllCombinedEquipments();
+        if ($scope.tabInitialized) {
+            $scope.getAllCombinedEquipments();
+        }
     });
 });