+++ myems-admin/app/controllers/settings/equipment/equipmentmeter.controller.js
@@ -217,13 +217,34 @@ app.controller('EquipmentMeterController', function(
         });
     };
 
-	$scope.getAllEquipments();
-	$scope.getAllMeters();
-	$scope.getAllVirtualMeters();
-	$scope.getAllOfflineMeters();
+	$scope.tabInitialized = false;
+
+	$scope.initTab = function() {
+		if (!$scope.tabInitialized) {
+			$scope.tabInitialized = true;
+			$scope.getAllEquipments();
+			$scope.getAllMeters();
+			$scope.getAllVirtualMeters();
+			$scope.getAllOfflineMeters();
+		}
+	};
+
+	$scope.$on('equipment.tabSelected', function(event, tabIndex) {
+		if ($scope.$parent && $scope.$parent.TAB_INDEXES && tabIndex === $scope.$parent.TAB_INDEXES.BIND_METER && !$scope.tabInitialized) {
+			$scope.initTab();
+		}
+	});
+
+	$timeout(function() {
+		if ($scope.$parent && $scope.$parent.TAB_INDEXES && $scope.$parent.activeTabIndex === $scope.$parent.TAB_INDEXES.BIND_METER && !$scope.tabInitialized) {
+			$scope.initTab();
+		}
+	}, 0);
 
   	$scope.$on('handleBroadcastEquipmentChanged', function(event) {
-    	$scope.getAllEquipments();
+		if ($scope.tabInitialized) {
+			$scope.getAllEquipments();
+		}
   	});
 });