+++ myems-admin/app/controllers/settings/meter/offlinemeterfile.controller.js
@@ -4,7 +4,8 @@ app.controller('OfflineMeterFileController', function(
 	$scope, 
 	$window,
 	$translate,
-	$interval, 
+	$interval,
+	$timeout,
 	OfflineMeterFileService, 
 	toaster, 
 	SweetAlert) {
@@ -113,16 +114,36 @@ app.controller('OfflineMeterFileController', function(
 			});
 	};
 
-	$scope.getAllOfflineMeterFiles();
-	$interval.cancel();
+	$scope.tabInitialized = false;
+	$scope.refeshfiles = undefined;
+
+	$scope.initTab = function() {
+		if (!$scope.tabInitialized) {
+			$scope.tabInitialized = true;
+			$scope.getAllOfflineMeterFiles();
+			$interval.cancel();
+			$scope.refeshfiles = $interval($scope.getAllOfflineMeterFiles, 1000 * 8);
+		}
+	};
+
+	$scope.$on('meter.tabSelected', function(event, tabIndex) {
+		if ($scope.$parent && $scope.$parent.TAB_INDEXES && tabIndex === $scope.$parent.TAB_INDEXES.OFFLINE_METER_FILE && !$scope.tabInitialized) {
+			$scope.initTab();
+		}
+	});
+
+	$timeout(function() {
+		if ($scope.$parent && $scope.$parent.TAB_INDEXES && $scope.$parent.activeTabIndex === $scope.$parent.TAB_INDEXES.OFFLINE_METER_FILE && !$scope.tabInitialized) {
+			$scope.initTab();
+		}
+	}, 0);
 
 	$scope.$on('$destroy', function() {
-	     // Make sure that the interval is destroyed too
-	     if (angular.isDefined($scope.refeshfiles)) {
-	         $interval.cancel($scope.refeshfiles);
-	         $scope.refeshfiles = undefined;
-	     }
+		// Make sure that the interval is destroyed too
+		if (angular.isDefined($scope.refeshfiles)) {
+			$interval.cancel($scope.refeshfiles);
+			$scope.refeshfiles = undefined;
+		}
 	});
-	$scope.refeshfiles=$interval($scope.getAllOfflineMeterFiles,1000*8);
 
 });