+++ CHANGELOG.md
@@ -7,6 +7,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ### Added
 ### Changed
 ### Fixed
+- optimized combined equipment settings in myems-admin
+- optimized equipment settings in myems-admin
+- optimized meter settings in myems-admin
 ### Removed
 
 ## [v5.12.0] - 2025-12-26